restructuring (move Header&Layouts to App)
