+++ src/App/Layouts/MainLayout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import { HeaderMobile, HeaderDeck } from '@components/Header';
-import { useMediaQuery } from '@utils/useMediaQuery';
+import { HeaderMobile, HeaderDeck } from '@/App/Header';
+import { useMediaQuery } from '@utils/useMediaQuery.ts';
 
 const MainLayout: React.FC = () => {
   const isMobile = useMediaQuery('(max-width: 768px)');
