+++ src/App/Layouts/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MainLayout.tsx';
